refactor: classroom/teacher_router.py → teacher_class_router.py 파일명 명확화
- 반(class) 관리 라우터임을 파일명에서 명시적으로 드러내기 위해 변경
- teacher_router라는 이름은 classroom 도메인 내 역할을 충분히 설명하지 못했다
