+++ tests/structure/test_domain_structure.py
@@ -15,7 +15,7 @@ def test_domain_packages_are_importable():
         "app.domains.auth.schemas",
         "app.domains.auth.service",
         "app.domains.classroom.schemas",
-        "app.domains.classroom.teacher_router",
+        "app.domains.classroom.teacher_class_router",
         "app.domains.progress.teacher_router",
         "app.domains.classroom.dependencies",
         "app.domains.classroom.models",
